# perf: use prepared statment cache