This PR introduces a cache for prepared statements which improves performance by ~10% at 100VUs

Typical load test results after pre-warming on my macbook:

```bash
k6 run --vus 100 --iterations 10000 test.js

master
    iteration_duration.........: avg=55.42ms min=15.67ms med=55.18ms max=90.3ms  p(90)=63.34ms p(95)=66.49ms
    iterations.................: 10000  1797.863942/s

this PR
    iteration_duration.........: avg=44.16ms min=13.45ms med=43.4ms  max=79.23ms p(90)=47.22ms p(95)=49.71ms
    iterations.................: 10000  2253.759616/s
```