@missinglink 2020-03-17
superseded by https://github.com/pelias/interpolation/pull/244